Fullpath to docker-compose
Fun fact: not including the full path to an executeable in a sudoers.d file will
result in a parse error that breaks sudo system wide if you save it.  To
recover, you need to mount the disk to a different system and correct the parse
error.  Not that I know from experience or anything.

PS. This is why you should always use `visudo` (or `visudo -f FILE`) whenever
you touch a file controlling access to sudo.